extract common object converter to io package
