+++ platform/src/main/java/dev/kkorolyov/pancake/platform/io/ObjectParser.java
@@ -1,4 +1,4 @@
-package dev.kkorolyov.pancake.platform.registry;
+package dev.kkorolyov.pancake.platform.io;
 
 import dev.kkorolyov.flub.function.convert.Converter;
